backend/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
bidict==0.22.1
bjson==0.2.9
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.96.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.1.2
orjson==3.9.0
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
PyJWT==1.7.1
pymongo==4.3.3
python-dotenv==1.0.0
python-engineio==4.4.1
python-multipart==0.0.6
python-socketio==5.8.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
